change types

+++ src/board.rs
@@ -34,14 +34,14 @@ pub fn point_to_algebraic(x: Point) -> String {
 
 pub fn algebraic_to_point(x: &str) -> Point {
     let c: Vec<char> = x.chars().collect();
-    let col_idx = col_chr_to_idx(c[0]) as i8;
+    let col_idx = col_chr_to_idx(c[0]) as isize;
     let n = c.len();
     let mut num_str = String::new();
     for i in 1..n {
         num_str.push(c[i]);
     }
     let row_num: usize = (&num_str).parse().unwrap();
-    let row_idx = row_num_to_idx(row_num) as i8;
+    let row_idx = row_num_to_idx(row_num) as isize;
     return (row_idx, col_idx);
 }
